refactor(bot): break Cycle C via direct imports + extract audioFeatures/vcWeights (#889) (#931)
Resolves #889 — last runtime circular dependency in
\`packages/bot/src\`.

## What changed

Cycle C was \`queueManipulation.ts ↔ autoplay/replenisher.ts\` via a
6-symbol barrel import. 4 of those symbols already lived in standalone
files (\`candidateFallback.ts\`); the other 2 needed minimal extraction:

- \`getTrackAudioFeatures\` + \`audioFeatureCache\` →
\`autoplay/audioFeatures.ts\`
- \`buildVcContributionWeights\` → \`autoplay/vcWeights.ts\`

\`replenisher.ts\` now imports each symbol directly from its source.
\`queueManipulation.ts\` and \`queueRescue.ts\` retain their public
surface via re-export, so no consumer touches.

## Madge

- Before: 2 cycles (Cycle C + deferred type-only \`types/CustomClient\`
cycle)
- After: 1 cycle (only the deferred type-only one remains, as documented
in #871)

Unblocks promotion of \`.github/workflows/madge.yml\` from
\`continue-on-error: true\` to a hard gate (separate follow-up PR).

## Tests

- 433/433 autoplay + queue + replenisher tests pass
- 3075/3075 total bot suite pass (2 unrelated \`prom-client\` env
failures in worktree only)

## Closure-capture avoidance

The prior agent attempt at Cycle C broke 107 tests due to
closure-captured state. This change keeps \`audioFeatureCache\` as a
module-level \`const LRUCache\` — purely lexical relocation, no shape
change.

## ADR


\`docs/decisions/2026-05-21-cycle-c-direct-imports-and-audio-features-extraction.md\`
records the decision, alternatives (B: larger extraction; C: accept the
cycle; D: callback indirection), and revisit triggers.

## Changelog

[Unreleased] Changed: break the last runtime circular dependency in
\`packages/bot/src\` (Cycle C, #889).
